Attempt to Create Table
Find that number of occurrences of certain table values is less than others. Start, but do not finish a script that duplicates single values to match the longest value list. Comment out progress.

Change the parameters of build_table_str to *args from **kwargs for ease of use

Generate a table string on every app run but do not insert it into file. Comment out.