Corrected Last-Row-First Behavior
Find an issue in populate_table where the -1 index is called.

This was caused by a prior commit removing the start-at-one rule within the range function because it resulted in empty tables.

To prevent this and keep rows in the proper order, the list now ranges from 1 to len(table) + 1.